+++ src/SeerUtl.cpp
@@ -7,7 +7,7 @@
 // Increment this with every release on GitHub.
 // See scripts/change_versionnumber
 //
-#define SEER_VERSION "1.13beta"
+#define SEER_VERSION "1.13"
 
 namespace Seer {